+++ android/app/src/main/java/io/coursepick/coursepick/view/LocationProvider.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick
+package io.coursepick.coursepick.view
 
 import android.Manifest
 import android.content.Context
